% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibiotics_metadata.R
\docType{data}
\name{antibiotics_metadata}
\alias{antibiotics_metadata}
\title{Antibiotics metadata for the Yassour et al. study}
\format{A data frame with 38 observations of 9 variables
\describe{
  \item{Subject}{Baby subject ID}
  \item{Age}{Age when antibiotic taken}
  \item{Antibiotic type}{Name of the antibiotic given}
  \item{Antibiotic code}{The antibiotic's ATC classification code}
  \item{Duration}{Length of time the antibiotic was taken}
  \item{Symptoms}{The reason for antibiotic prescription}
  
}}
\source{
\url{https://pubs.broadinstitute.org/diabimmune/antibiotics-cohort}
}
\usage{
antibiotics_metadata
}
\description{
Summarized metadata describing antibiotic exposure for the 38 infants in this cohort.
}
\keyword{datasets}
