% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_otu_table.R
\docType{data}
\name{filtered_otu_table}
\alias{filtered_otu_table}
\title{OTU table from Yassour et al. microbiome study}
\format{A data frame with 1101 samples (columns) and 3091 taxa (rows):
\describe{
  \item{samples}{Each baby ID in format E999999, plus the timepoint in months}
  \item{OTU_ID}{Relative abundance of taxa, summarized at each level of taxonomy}
  
}}
\source{
\url{https://pubs.broadinstitute.org/diabimmune/antibiotics-cohort}
}
\usage{
filtered_otu_table
}
\description{
The 16S OTU table from Yassour et al. (Science Trans Med, 2016), generated with
QIIME v1.8.0 and picked with Greengenes (October 2012).
}
\keyword{datasets}
