% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_metadata.R
\docType{data}
\name{general_metadata}
\alias{general_metadata}
\title{General metadata for the Yassour et al. study}
\format{A data frame with 38 observations of 9 variables
\describe{
  \item{birth_year}{year born in YYYY}
  \item{sex}{baby's sex, male or female}
  \item{birth_mode}{was the baby born by C-section or vaginally?}
  \item{gestational_age}{weeks into pregnancy as weeks+days}
  \item{months_final_exclusive_breastfeeding}{when did exclusive breast feeding end?}
  \item{months_end_breastfeeding}{when did any/all breastfeeding end}
  \item{months_started_formula_feeding}{when was formula introduced, if ever}
  \item{feeding_in_hospital}{type of feeding in hospital}
  
}}
\source{
\url{https://pubs.broadinstitute.org/diabimmune/antibiotics-cohort}
}
\usage{
general_metadata
}
\description{
Metadata for the 38 infants including information about age, sex, feeding status, and
birth mode.
}
\keyword{datasets}
