% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_permusplinectomy.R
\name{permuspliner.plot.permsplines}
\alias{permuspliner.plot.permsplines}
\title{Plot permuted splines along the real data}
\usage{
permuspliner.plot.permsplines(data, xvar = NULL, yvar = NULL)
}
\arguments{
\item{data}{The results object from the permuspliner function}

\item{xvar}{Name (as string) of the longitudinal x variable in the data}

\item{yvar}{Name (as string) of the response/y variable in the data}
}
\description{
Compare how the permuted splines fit with the real data. Provides visual support for p values.
}
\examples{
permsplot <- permuspliner.plot.permsplines(results, xvar='Time', yvar='Weight')

# View the plot
permsplot

# Save the plot as PNG file
ggsave(permsplot, file = 'my_plot.png', dpi=300, height=4, width=4)

}
