% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_sliding_spline.R
\name{sliding_spliner.plot.pvals}
\alias{sliding_spliner.plot.pvals}
\title{Plot sliding spliner p-values}
\usage{
sliding_spliner.plot.pvals(data, xvar = "xvar")
}
\arguments{
\item{data}{The dataframe of p-values over the x variable}

\item{xvar}{The independent variable; is continuous, e.g. time.}
}
\description{
Plot the imputed p-values over the x variable range
}
\examples{
sliding_spliner.plot.pvals(result$pval_table,
 xvar = 'Time')

}
