% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_trendyspline.R
\name{trendyspliner.plot.perms}
\alias{trendyspliner.plot.perms}
\title{Plot permuted trends behind the real data}
\usage{
trendyspliner.plot.perms(data = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{data}{Required: The results object from the trendyspliner function}

\item{xlabel}{Optional: Title (as string) to print on the x axis}

\item{ylabel}{Optional: Title (as string) to print on the y axis}
}
\description{
Compare how the permuted trend splines fit with the real data. Provides visual support for p values.
}
\examples{

# Loads ggplot2 dependency
permsplot <- trendyspliner.plot.perms(results, xlabel='Time', ylabel='Weight')

# View the plot
permsplot

# Save the plot as PNG file
ggsave(permsplot, file = 'my_plot.png', dpi=300, height=4, width=4)

}
