% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_trendyspline.R
\name{trendyspliner}
\alias{trendyspliner}
\title{Trendyspliner}
\usage{
Permutation to test whether there is a non-zero trend among a set
of individuals/samples over a continuous variable -- such as time. So, there
does not need to be two groups in this test. The x variable datapoints are
permuated within each case/individual, thus maintaining the distribution in
the y component but shuffling the hypothesized trend.
}
\arguments{
\item{data}{A dataframe object containing your data.}

\item{category}{The column name of the category to be tested, if present.}

\item{cases}{The column name defining the individual cases, e.g. patients.}

\item{group}{If more than two groups, the two groups to compare.}

\item{perms}{The number of permutations to generate}

\item{set_spar}{Set the spar parameter for splines}

\item{cut_low}{Remove data with fewer than __ points}

\item{ints}{Number of x intervals over which to measure area}

\item{quiet}{Silence all text outputs}

\item{x}{The independent variable; is continuous, e.g. time.}

\item{y}{The dependent variable; is continuous, e.g. temperature.}
}
\description{
Test for a significant non-zero trend in a response over time.
}
\details{
The data object needs to be organized with each observation as a
row, and have a column that identifies the case, patient, animal, etc, and
columns with the continuous x and y variables (row with x = NA will be
removed). If there are multiple groups in the data, you can filter to
the single group of interest with the category and group arguments. Otherwise
it assumes the entire dataset is the single population.
}
\examples{
trendyspliner(data = ChickWeight, xvar = 'Time',
             yvar = 'weight', category = 'Diet',
             cases = 'Chick', group = '1')
}
