% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_permusplinectomy.R
\name{permuspliner}
\alias{permuspliner}
\title{Permuspliner}
\usage{
Permutation to test whether there is a non-zero trend among a set
of individuals/samples over a continuous variable -- such as time. So, there
does not need to be two groups in this test. The x variable datapoints are
permuated within each case/individual, thus maintaining the distribution in
the y component but shuffling the hypothesized trend.
}
\arguments{
\item{data}{A dataframe object containing your data.}

\item{category}{The column name of the category to be tested.}

\item{cases}{The column name defining the individual cases, e.g. patients.}

\item{groups}{If more than two groups, the two groups to compare.}

\item{perms}{The number of permutations to generate}

\item{set_spar}{Set the spar parameter for splines}

\item{cut_low}{Remove individual cases with fewer than _ observations}

\item{ints}{Number of x intervals over which to measure area}

\item{x}{The independent variable; is continuous, e.g. time.}

\item{y}{The dependent variable; is continuous, e.g. temperature.}
}
\description{
Tests for a significant difference between two groups overall.
}
\examples{
permuspliner(data = ChickWeight, xvar = 'Time',
             yvar = 'weight', category = 'Diet',
             cases = 'Chick', groups = '1,4')
}
