% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_sliding_spline.R
\name{sliding_spliner.plot.pvals}
\alias{sliding_spliner.plot.pvals}
\title{Plot sliding spliner p-values}
\usage{
sliding_spliner.plot.pvals(data_df, cases = cases, xvar = xvar)
}
\arguments{
\item{cases}{The column name defining the individual cases, e.g. patients.}

\item{data}{The dataframe of p-values over the x variable}

\item{x}{The independent variable; is continuous, e.g. time.}
}
\description{
Plot the imputed p-values over the x variable range
}
