% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_sliding_spline.R
\name{sliding_spliner.plot.splines}
\alias{sliding_spliner.plot.splines}
\title{Plot group splines}
\usage{
sliding_spliner.plot.splines(data, category = "category", xvar = "xvar",
  yvar = "yvar")
}
\arguments{
\item{data}{The dataframe of imputed points generated from splines}

\item{category}{The data category being compared}

\item{xvar}{The independent variable; is continuous, e.g. time.}

\item{yvar}{The dependent variable; is continuous, e.g. temperature.}
}
\description{
Plot the individual splines grouped by two categories of interest
}
\examples{
sliding_spliner.plot.splines(result,
 category = 'Diet', xvar = 'Time', yvar = 'weight')

}
